[REMANIEMENT] Délace chaque événement dans son fichier
